+++ instruments/tape/src/instrument/attributes.rs
@@ -1,1 +1,4 @@
 pub const PRE_GAIN: &str = "pre gain";
+pub const DRIVE: &str = "drive";
+pub const SATURATION: &str = "saturation";
+pub const WIDTH: &str = "width";