+++ instruments/tape/src/instrument/command.rs
@@ -4,9 +4,13 @@ use proton_ui::reaction::Reaction;
 
 use super::attributes;
 
+#[allow(clippy::enum_variant_names)]
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Command {
     SetPreGain(f32),
+    SetDrive(f32),
+    SetSaturation(f32),
+    SetWidth(f32),
 }
 
 use Command::*;
@@ -18,6 +22,9 @@ impl TryFrom<Reaction> for Command {
         match other {
             Reaction::SetValue(name, value) => match name {
                 attributes::PRE_GAIN => Ok(SetPreGain(value)),
+                attributes::DRIVE => Ok(SetDrive(value)),
+                attributes::SATURATION => Ok(SetSaturation(value)),
+                attributes::WIDTH => Ok(SetWidth(value)),
                 _ => Err("invalid attribute name"),
             },
             Reaction::SelectValue(_, _) => Err("invalid attribute type"),