+++ control/src/input_snapshot.rs
@@ -1,0 +1,11 @@
+#[derive(Clone, Copy, Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct InputSnapshot {
+    pub cv: [Cv; 5],
+}
+
+#[derive(Clone, Copy, Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct Cv {
+    pub value: f32,
+}