+++ peripherals/src/lib.rs
@@ -1,6 +1,7 @@
 #![no_std]
 
 pub mod button;
+pub mod cv_input;
 pub mod rotary;
 
 mod debounce_buffer;